Add Mock Backends Repair permissions bug
