+++ payment/payment_backends/__init__.py
@@ -1,3 +1,5 @@
 from .base import BaseBackend
+from .mock import MockPaymentBackend, AutoPayMockBackend
 from .nextpay import NextpayBackend
+from .request import BaseRequestBackend
 from .zibal import ZibalBackend
